requirements.txt
# Core dependencies
elasticsearch>=8.0.0
sentence-transformers>=2.2.0
numpy>=1.21.0
torch>=1.9.0
transformers>=4.30.0

# Environment and configuration
python-dotenv>=0.19.0
pyyaml>=6.0.0

# Utilities
tqdm>=4.65.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Logging and monitoring
prometheus-client>=0.16.0
python-json-logger>=2.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Development
black>=22.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0